+++ Xplat/src/main/java/vazkii/botania/common/block/corporea/CorporeaCrystalCubeBlock.java
@@ -45,8 +45,7 @@ public class CorporeaCrystalCubeBlock extends BotaniaWaterloggedBlock implements
 
 	@Override
 	public void attack(BlockState state, Level world, BlockPos pos, Player player) {
-		if (!world.isClientSide) {
-			CorporeaCrystalCubeBlockEntity cube = (CorporeaCrystalCubeBlockEntity) world.getBlockEntity(pos);
+		if (!world.isClientSide && world.getBlockEntity(pos) instanceof CorporeaCrystalCubeBlockEntity cube) {
 			cube.doRequest(player);
 		}
 	}
@@ -64,7 +63,9 @@ public class CorporeaCrystalCubeBlock extends BotaniaWaterloggedBlock implements
 			if (stack.getItem() instanceof WandOfTheForestItem && player.isSecondaryUseActive()) {
 				return InteractionResult.PASS;
 			}
-			CorporeaCrystalCubeBlockEntity cube = (CorporeaCrystalCubeBlockEntity) world.getBlockEntity(pos);
+			if (!(world.getBlockEntity(pos) instanceof CorporeaCrystalCubeBlockEntity cube)) {
+				return InteractionResult.FAIL;
+			}
 			if (cube.locked) {
 				if (!world.isClientSide) {
 					player.displayClientMessage(Component.translatable("botaniamisc.crystalCubeLocked"), false);
@@ -99,6 +100,8 @@ public class CorporeaCrystalCubeBlock extends BotaniaWaterloggedBlock implements
 
 	@Override
 	public int getAnalogOutputSignal(BlockState state, Level world, BlockPos pos) {
-		return ((CorporeaCrystalCubeBlockEntity) world.getBlockEntity(pos)).getComparatorValue();
+		return world.getBlockEntity(pos) instanceof CorporeaCrystalCubeBlockEntity cube
+				? cube.getComparatorValue()
+				: 0;
 	}
 }