+++ Xplat/src/main/java/vazkii/botania/common/item/FloralObedienceStickItem.java
@@ -13,7 +13,6 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -51,8 +50,8 @@ public class FloralObedienceStickItem extends Item {
 					continue;
 				}
 
-				BlockEntity tile = world.getBlockEntity(iterPos);
-				if (tile instanceof BindableSpecialFlowerBlockEntity<?>bindable && bindable.wouldBeValidBinding(pos)) {
+				if (world.getBlockEntity(iterPos) instanceof BindableSpecialFlowerBlockEntity<?>bindable
+						&& bindable.wouldBeValidBinding(pos)) {
 					bindable.setBindingPos(pos);
 					WandOfTheForestItem.doParticleBeamWithOffset(world, iterPos, pos);
 				}
@@ -60,8 +59,7 @@ public class FloralObedienceStickItem extends Item {
 
 			return true;
 		}
-		BlockEntity tile = world.getBlockEntity(pos);
-		if (tile instanceof BindableSpecialFlowerBlockEntity<?>bindableFlower) {
+		if (world.getBlockEntity(pos) instanceof BindableSpecialFlowerBlockEntity<?>bindableFlower) {
 			bindableFlower.setBindingPos(null);
 			return true;
 		}